+++ go.mod
@@ -6,22 +6,22 @@ require (
 	github.com/bwmarrin/snowflake v0.3.0
 	github.com/gin-gonic/gin v1.6.3
 	github.com/go-playground/assert/v2 v2.0.1
-	github.com/go-resty/resty/v2 v2.6.0
-	github.com/go-sql-driver/mysql v1.5.0
+	github.com/go-resty/resty/v2 v2.7.0
+	github.com/go-sql-driver/mysql v1.6.0
 	github.com/grpc-ecosystem/go-grpc-middleware v1.3.0
 	github.com/lib/pq v1.10.3
 	github.com/prometheus/client_golang v1.11.0
 	github.com/stretchr/testify v1.7.0
+	github.com/yedf/dtmdriver v0.0.0-20211202092330-86dce681d5db
+	github.com/yedf/dtmdriver-gozero v0.0.0-20211202141526-9fe6083e31d2
+	github.com/yedf/dtmdriver-protocol1 v0.0.0-20211202134026-9ae13beea958
+	go.uber.org/automaxprocs v1.4.1-0.20210525221652-0180b04c18a7
 	golang.org/x/crypto v0.0.0-20211108221036-ceb1ce70b4fa // indirect
-	golang.org/x/net v0.0.0-20210805182204-aaa1db679c0d // indirect
-	golang.org/x/sys v0.0.0-20210806184541-e5e7981a1069 // indirect
-	google.golang.org/genproto v0.0.0-20210805201207-89edb61ffb67 // indirect
-	google.golang.org/grpc v1.39.1
+	google.golang.org/grpc v1.42.0
 	google.golang.org/protobuf v1.27.1
-	gopkg.in/yaml.v2 v2.3.0
+	gopkg.in/yaml.v2 v2.4.0
 	gorm.io/driver/mysql v1.0.3
 	gorm.io/driver/postgres v1.2.1
 	gorm.io/gorm v1.22.2
-	go.uber.org/automaxprocs v1.4.1-0.20210525221652-0180b04c18a7
 // gotest.tools v2.2.0+incompatible
 )
